feat: detailinfo 평점 커스텀훅 분리
