+++ src/components/DetailInfo/DetailInfo.jsx
@@ -1,5 +1,5 @@
-import React, { useEffect } from 'react';
-import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query';
+import React from 'react';
+import { useGetAverageRating } from '../../hooks/useGetAverageRating';
 import {
   DetailInfoWrapper,
   InfoItem,
@@ -8,53 +8,9 @@ import {
   Star,
   InfoContainer
 } from '../../styles/Detail/DetailInfo/detailInfoStyle';
-import supabase from '../../supabase/supabaseClient';
-
-const fetchAverageRating = async (shopId) => {
-  const { data, error } = await supabase.from('comments').select('rating').eq('shopId', shopId);
-
-  if (error) {
-    throw new Error(error.message);
-  }
-
-  if (data && data.length > 0) {
-    const totalRating = data.reduce((acc, comment) => acc + comment.rating, 0);
-    const avgRating = totalRating / data.length;
-    return avgRating;
-  } else {
-    return null; // 리뷰가 없을 때 null로 설정
-  }
-};
-
-const updateComment = async ({ shopId, commentId, rating }) => {
-  const { data, error } = await supabase.from('comments').update({ rating }).eq('id', commentId).eq('shopId', shopId);
-
-  if (error) {
-    throw new Error(error.message);
-  }
-
-  console.log(data);
-  return data;
-};
-
-export const useUpdateComment = (shopId) => {
-  const queryClient = useQueryClient();
-
-  return useMutation({
-    mutationFn: updateComment,
-    onSuccess: () => queryClient.invalidateQueries({ queryKey: ['averageRating', shopId] })
-  });
-};
 
 const DetailInfo = ({ shop }) => {
-  const {
-    data: rating,
-    error,
-    isLoading
-  } = useQuery({
-    queryKey: ['averageRating', shop.id],
-    queryFn: () => fetchAverageRating(shop.id)
-  });
+  const { data: rating, error, isLoading } = useGetAverageRating(shop.id);
 
   const renderStars = (rating) => {
     const stars = [];
